requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.16.0
Django==5.0.1
django-phonenumber-field==7.3.0
django-rest==0.8.7
djangorestframework==3.14.0
phonenumbers==8.13.28
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
